+++ neodos-kernel/src/shell/commands/tsr.rs
@@ -1,8 +1,0 @@
-use crate::println;
-use crate::shell::shell::DosShell;
-
-impl DosShell {
-    pub fn cmd_tsr(&mut self, _args: &[&str]) {
-        println!("TSR not implemented in VFS shell yet");
-    }
-}